{"schema":"decodex/commit/1","summary":"Move backend image capture helpers","authority":"manual"}
